Add files via upload
This is the current state of my DCF model. I am continuing to work on it as well as complete the other components of an equity research report. This is for educational purposed only. All data within is from public records (companies own annual report and industry publications about core assumptions made)